requirements.txt
kagglehub
pandas
python-dotenv
argparse
matplotlib